handler and router files swapped Relates to #33
