+++ src/server.js
@@ -1,8 +1,8 @@
 const http = require("http");
 
-const handler = require("./handler.js");
+const router = require("./router.js");
 
-const server = http.createServer(handler);
+const server = http.createServer(router);
 
 const port = 3000;
